refactor: KISS configureStore
