+++ src/drive/mobile/lib/background.js
@@ -1,7 +1,7 @@
-import configureStore from '../../store/configureStore'
+import configureStore from '../store/configureStore'
 import { initServices } from './init'
 import { logException, logInfo } from './reporter'
-import { loadState } from '../../store/persistedState'
+import { loadState } from '../store/persistedState'
 import { getMediaFolderName } from './media'
 import { startMediaBackup } from '../actions/mediaBackup'
 import { isCordova, isIos, isAndroid, getPlatformId } from './device'
