+++ src/drive/store/configureStore.js
@@ -3,39 +3,28 @@ import { compose, createStore, applyMiddleware } from 'redux'
 import { shouldEnableTracking, getTracker, createTrackerMiddleware } from 'cozy-ui/react/helpers/tracker'
 import thunkMiddleware from 'redux-thunk'
 import eventTrackerMiddleware from '../middlewares/EventTracker'
-import { getReducers as getTargetReducers, getMiddlewares as getTargetMiddlewares } from './getTargetConfig'
+import rootReducer from '../reducers'
+import logger from 'redux-logger'
 
-const getReducers = (persistedState) => {
-  let reducers = [getTargetReducers]
-
-  if (persistedState) {
-    reducers.push(persistedState)
-  }
-
-  return reducers
-}
-
-const getMiddlewares = () => {
-  let middlewares = [thunkMiddleware]
-
+const configureStore = (initialState = {}) => {
+  const middlewares = [
+    thunkMiddleware
+  ]
   if (shouldEnableTracking() && getTracker()) {
     middlewares.push(eventTrackerMiddleware)
     middlewares.push(createTrackerMiddleware())
   }
-
-  return middlewares.concat(getTargetMiddlewares())
-}
-
-const configureStore = (persistedState) => {
-  const reducers = getReducers(persistedState)
-  const middlewares = getMiddlewares()
+  if (__DEVELOPMENT__) {
+    middlewares.push(logger)
+  }
 
   // Enable Redux dev tools
   const composeEnhancers = (__DEVELOPMENT__ && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose
 
   const store = createStore(
-    ...reducers,
-    composeEnhancers(applyMiddleware.apply(this, middlewares))
+    rootReducer,
+    initialState,
+    composeEnhancers(applyMiddleware(...middlewares))
   )
 
   return store
