--- src/drive/store/index.js
@@ -1,5 +1,0 @@
-import configureStore from './configureStore'
-import { loadState } from './persistedState'
-import { persistState } from './getTargetConfig'
-
-export { configureStore, loadState, persistState }
